requirements.txt
streamlit
streamlit-autorefresh
paho-mqtt
RPi.GPIO
gpiozero
matplotlib
pandas
